refactor(input): extract OverlayState and handle_overlay_key to src/input/overlay.rs (partial #66)
Phase 0 of the IME overlay roadmap (Issue #82). Behavior is unchanged —
OverlayState struct, its buffer/cursor helpers, and the modal
handle_overlay_key routine are lifted out of src/app.rs into a new
src/input/overlay.rs module. App still owns Option<OverlayState> and the
open/close call sites, so no public API shifts and no tests needed
updating.

- Add `mod input;` with `pub mod overlay;` first slice.
- Re-export OverlayState via `pub use crate::input::overlay::OverlayState`
  so `crate::app::OverlayState` keeps resolving for call sites.
- Promote App::always_dismissed_pane to pub(crate) so the modal handler
  can still record explicit Always-mode dismissals.
- Call site in handle_key now dispatches to
  crate::input::overlay::handle_overlay_key(self, key).

This sets up the natural location for the pane-snapshot field Phase 2
(#37) will add, and unblocks Phase 1 (#38 throttle) on the same epic.

cargo check / clippy / test / fmt all green (237 tests pass).
